# packages/event/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
			"FormSubmission": {
				"properties": {
					"label": {
						"type": "string"
					},
					"fieldType": {
						"type": "string"
					},
					"answers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"label",
					"fieldType",
					"answers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EventDetail": {
				"properties": {
					"orgId": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"thumbnail": {
						"type": "string"
					},
					"eventName": {
						"type": "string"
					},
					"address": {
						"properties": {
							"lng": {
								"type": "string"
							},
							"lat": {
								"type": "string"
							}
						},
						"required": [
							"lng",
							"lat"
						],
						"type": "object"
					},
					"category": {
						"type": "string",
						"enum": [
							"Sport",
							"Education",
							"Workshop",
							"Charity"
						]
					},
					"viewer": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"Date": {
						"properties": {
							"endTime": {
								"type": "string"
							},
							"startTime": {
								"type": "string"
							},
							"endDate": {
								"type": "string",
								"format": "date-time"
							},
							"startDate": {
								"type": "string",
								"format": "date-time"
							}
						},
						"type": "object"
					},
					"requirement": {
						"properties": {
							"timeCommitment": {
								"type": "string"
							},
							"skill": {
								"type": "string"
							},
							"language": {
								"type": "string"
							},
							"age": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"formSubmission": {
						"items": {
							"$ref": "#/components/schemas/FormSubmission"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "event",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/v1/events/name": {
			"get": {
				"operationId": "FindEventByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/events/cate": {
			"get": {
				"operationId": "FindEventByCate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/events": {
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EventDetail"
							}
						}
					}
				}
			}
		},
		"/v1/events/{id}": {
			"put": {
				"operationId": "UpdateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EventDetail"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "FindFavoEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}